web-stack/services/hydro-toolkit/pyproject.toml
[project]
name = "hydro-toolkit"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
]
